refactor(Rank): 패키지 constant로 변경
