+++ src/test/java/lotto/domain/PickerTest.java
@@ -9,6 +9,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import lotto.constant.Rank;
+
 class PickerTest {
     private List<Lotto> userLottos;
